Replace most property contents with calls to new SDS011._rw_property function
The _rw_property writes returned values from the device to memory so the last known value may be returned if the device is sleeping. This commit does not include the device_id property, which still returns the "all devices" id, and still returns it as a raw bytestring instead of response dict.